+++ src/lib/widgets/book_card.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:cached_network_image/cached_network_image.dart';
 import '../models/book.dart';
+import '../config/app_colors.dart';
 
 class BookCard extends StatelessWidget {
   final Book book;
@@ -43,10 +44,10 @@ class BookCard extends StatelessWidget {
                       ? CachedNetworkImage(
                           imageUrl: book.coverUrl!,
                           fit: BoxFit.cover,
-                          placeholder: (_, __) => _placeholder(),
-                          errorWidget: (_, __, ___) => _placeholder(),
+                          placeholder: (_, __) => _placeholder(context),
+                          errorWidget: (_, __, ___) => _placeholder(context),
                         )
-                      : _placeholder(),
+                      : _placeholder(context),
                 ),
               ),
               const SizedBox(width: 12),
@@ -160,8 +161,8 @@ class BookCard extends StatelessWidget {
     );
   }
 
-  Widget _placeholder() => Container(
-        color: const Color(0xFFD6EAF8),
+  Widget _placeholder(BuildContext ctx) => Container(
+        color: AppColors.chipBg(ctx),
         child: const Center(
           child: Icon(Icons.menu_book, color: Color(0xFF1A5276), size: 28),
         ),